fix codeql circular import betweeen accelerator and system_config
